refactor: Rename NotificationMessenger to SystemNotificationMessenger
