+++ src/aux-records/index.ts
@@ -50,7 +50,7 @@ export * from './ModerationController';
 export * from './ModerationStore';
 export * from './ModerationConfiguration';
 export * from './ModerationJobProvider';
-export * from './NotificationMessenger';
+export * from './SystemNotificationMessenger';
 
 export * from './LoomController';
